Refactor namespaces and improve repository methods
Updated namespaces in `IOrderStatusRepository.cs` and `OrderStatusRepository.cs` to align with project structure. Refactored `CreateAsync` in `SaleTransactionRepository.cs` to use `AddAsync` for proper async behavior. Adjusted `DeleteAsync` to update and return the correct entity.
