+++ Dentizone.Application/Interfaces/IOrderStatusRepository.cs
@@ -1,7 +1,7 @@
 ﻿using Dentizone.Application.Interfaces;
 using Dentizone.Domain.Entity;
 
-namespace Dentizone.Application.Repositories;
+namespace Dentizone.Application.Interfaces;
 
 interface IOrderStatusRepository : IBaseRepo<OrderStatus>
 {
