+++ Dentizone.Application/Repositories/SaleTransactionRepository.cs
@@ -15,26 +15,26 @@ public class SaleTransactionRepository : AbstractRepository, ISaleTransactionRep
     public async Task<SalesTransaction?> GetByIdAsync(string id)
     {
         return await dbContext.SalesTransactions
-                              .Include(st => st.Wallet)
-                              .Include(st => st.Payment)
-                              .FirstOrDefaultAsync(st => st.Id == id && !st.IsDeleted);
+            .Include(st => st.Wallet)
+            .Include(st => st.Payment)
+            .FirstOrDefaultAsync(st => st.Id == id && !st.IsDeleted);
     }
 
     public async Task<IEnumerable<SalesTransaction>> GetAllAsync(int page = 1)
     {
         int skip = CalculatePagination(page);
         return await dbContext.SalesTransactions
-                              .Include(st => st.Wallet)
-                              .Include(st => st.Payment)
-                              .Where(st => !st.IsDeleted)
-                              .Skip(skip)
-                              .Take(DefaultPageSize)
-                              .ToListAsync();
+            .Include(st => st.Wallet)
+            .Include(st => st.Payment)
+            .Where(st => !st.IsDeleted)
+            .Skip(skip)
+            .Take(DefaultPageSize)
+            .ToListAsync();
     }
 
     public async Task<SalesTransaction> CreateAsync(SalesTransaction entity)
     {
-        dbContext.SalesTransactions.Add(entity);
+        await dbContext.SalesTransactions.AddAsync(entity);
         await dbContext.SaveChangesAsync();
         return entity;
     }
@@ -60,8 +60,8 @@ public class SaleTransactionRepository : AbstractRepository, ISaleTransactionRep
             return null;
         }
 
-        dbContext.SalesTransactions.Update(existingEntity);
+        dbContext.SalesTransactions.Update(entity);
         await dbContext.SaveChangesAsync();
-        return existingEntity;
+        return entity;
     }
 }
